Geolocation/Permissions Checks/Temp XML text input and buttons
Added geolocation that takes in text input from a user and converts into an array that contains longitude and latitude.
Added permissions check for google map.
Added button that geolocates on click.
Added button that goes to user location if permissions are enabled. If the position is not enabled the app will ask for permissions again.