set-markers in a statically coded list
Set up the code to clear old markers and set new ones every time the map view changes with search or jump to the current location button. 
The getPlaces method just sets the premade list of lists to itself but needs to be worked on to retrieve the list from the DB. If a properly formatted list is returned the rest of the code should work and if not some minor adjustments may be needed.