Clean MapFragment
Clean up the map fragment to have more consistency in the code.
Issue: If location is disabled must hit the search button twice to get the map markers to show so I defaulted it to show locations at the default view of the map but it appears the permissions check is stopping the marker placement on load. However, the search function still places the marks.